requirements.txt
pydantic>=2.0,<3.0
pyyaml>=6.0.1
gradio>=4.44.0
fastapi>=0.110.0
uvicorn>=0.29.0
openai>=1.0.0
hypothesis>=6.100.0
pytest>=8.0.0
